+++ core/src/i18n/en.yaml
@@ -1973,6 +1973,8 @@ virtualservices:
   tls-routes:
     matches: Matches
     title: TLS Route
+workload-selector:
+  title: Workload Selector
 workloads:
   docker-image:
     tooltip: Enter a valid Docker image registry path.
